refactor: rename `server.historyApiFallback` to support preview server
