+++ .changeset/mean-jars-sin.md
@@ -1,0 +1,6 @@
+---
+'@rsbuild/shared': patch
+'@rsbuild/core': patch
+---
+
+refactor: rename `server.historyApiFallback` to support preview server
