+++ cspell.json
@@ -39,6 +39,7 @@
     "selfsign",
     "sirv",
     "stacktracey",
+    "subpage",
     "svgr",
     "systemjs",
     "transpiling",
