+++ e2e/cases/server/basic/tsconfig.json
