+++ e2e/cases/server/history-api-fallback/src/index.tsx
