+++ e2e/cases/server/history-api-fallback/tsconfig.json
