+++ e2e/cases/server/hmr/src/App.tsx
