+++ e2e/cases/server/hmr/tsconfig.json
