+++ packages/document/docs/en/config/options/dev.mdx
@@ -32,12 +32,6 @@ import DevMiddleware from '@en/shared/config/dev/devMiddleware.md';
 
 <DevMiddleware />
 
-## dev.historyApiFallback
-
-import HistoryApiFallback from '@en/shared/config/dev/historyApiFallback.md';
-
-<HistoryApiFallback />
-
 ## dev.hmr
 
 import Hmr from '@en/shared/config/dev/hmr.md';
