+++ packages/document/docs/en/config/options/server.mdx
@@ -8,6 +8,12 @@ import Headers from '@en/shared/config/server/headers.md';
 
 <Headers />
 
+## server.historyApiFallback
+
+import HistoryApiFallback from '@en/shared/config/server/historyApiFallback.md';
+
+<HistoryApiFallback />
+
 ## server.htmlFallback
 
 import HtmlFallback from '@en/shared/config/server/htmlFallback.md';
