+++ packages/document/docs/zh/config/options/dev.mdx
@@ -32,12 +32,6 @@ import DevMiddleware from '@zh/shared/config/dev/devMiddleware.md';
 
 <DevMiddleware />
 
-## dev.historyApiFallback
-
-import HistoryApiFallback from '@zh/shared/config/dev/historyApiFallback.md';
-
-<HistoryApiFallback />
-
 ## dev.hmr
 
 import Hmr from '@zh/shared/config/dev/hmr.md';
