+++ packages/document/docs/zh/config/options/server.mdx
@@ -8,6 +8,12 @@ import Headers from '@zh/shared/config/server/headers.md';
 
 <Headers />
 
+## server.historyApiFallback
+
+import HistoryApiFallback from '@zh/shared/config/server/historyApiFallback.md';
+
+<HistoryApiFallback />
+
 ## server.htmlFallback
 
 import HtmlFallback from '@zh/shared/config/server/htmlFallback.md';
